# specs/003-add-a-matching/contracts/api.yaml
openapi: 3.0.3
info:
  title: Admin Dashboard API
  version: 1.0.0
  description: |
    API contracts for the admin dashboard content management system.
    Extends existing blog and portfolio APIs with admin-specific operations.
  contact:
    name: Rohan Batra
    email: contact@rohanbatra.com
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://rohanbatra.com/api
    description: Production server
security:
  - ClerkAuth: []
paths:
...
  /admin/users:
    get:
      summary: Get user list with admin data
      tags: [Admin User Management]
      security:
        - ClerkAuth: []
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [user, editor, admin]
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [createdAt, updatedAt, name, email]
            default: createdAt
      responses:
        '200':
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/AdminUser'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
...
components:
  securitySchemes:
    ClerkAuth:
      type: http
      scheme: bearer
      description: Clerk JWT token in Authorization header
  schemas:
...
    AdminUser:
      type: object
      properties:
        _id:
          type: string
        clerkId:
          type: string
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [user, editor, admin]
        avatar:
          type: string
        bio:
          type: string
        preferences:
          type: object
        activity:
          type: object
          properties:
            commentCount:
              type: integer
            likeCount:
              type: integer
            lastActive:
              type: string
              format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
...